research journey/research_journey_ai/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
google-generativeai==0.3.2
requests==2.31.0
python-dotenv==1.0.0
jinja2==3.1.2
sqlalchemy==2.0.23
pydantic==2.5.0
python-multipart==0.0.6
